fixed bug & renamed my functions.py to my_functions.py
